Teapot: Version Update (#4)
* Teapot: Cpp Version

- Cpp 17 to 20
- Cpp 20 features entegrated

* Teapot: OpenGL version

- 3.3 to 4.6
- GLAD update

+++ Teapot/src/Renderer/FrameBuffer.h
@@ -8,10 +8,10 @@ namespace Teapot
 	class FrameBuffer
 	{
 	public:
-		FrameBuffer(float width, float height);
+		FrameBuffer(int width, int height);
 		~FrameBuffer();
 		unsigned int GetFrameTexture() const;
-		void RescaleFrameBuffer(float width, float height) const;
+		void RescaleFrameBuffer(int width, int height) const;
 		void Bind() const;
 		void Unbind() const;
 	private:
